Refactor: study api, study-list api 분리
